+++ main.go
@@ -45,7 +45,7 @@ func run() error {
 	flag.StringVar(&g.address, "addr", ":8080", "The address to listen to")
 	flag.IntVar(&g.minDuration, "duration-min", 1, "Minimum request duration")
 	flag.IntVar(&g.maxDuration, "duration-max", 10, "Maximum request duration")
-	flag.IntVar(&g.errorsPercentage, "errors-percentage", 10, "Which percentage of the requests will fail")
+	flag.Float64Var(&g.errorsPercentage, "errors-percentage", 10, "Which percentage of the requests will fail")
 	flag.Parse()
 
 	return g.run()
@@ -55,7 +55,7 @@ type metricsGenerator struct {
 	address          string
 	minDuration      int
 	maxDuration      int
-	errorsPercentage int
+	errorsPercentage float64
 }
 
 func (g *metricsGenerator) run() error {
